+++ crates/warpgate/src/lib.rs
@@ -1,3 +1,4 @@
+mod client;
 mod endpoints;
 mod error;
 mod helpers;
@@ -6,6 +7,7 @@ mod loader;
 mod locator;
 mod plugin;
 
+pub use client::*;
 pub use error::*;
 pub use helpers::*;
 pub use id::*;