+++ crates/warpgate/tests/loader_test.rs
@@ -1,6 +1,6 @@
 use starbase_sandbox::{create_empty_sandbox, locate_fixture, Sandbox};
 use std::path::PathBuf;
-use warpgate::{GitHubLocator, Id, PluginLoader, PluginLocator};
+use warpgate::{create_http_client, GitHubLocator, Id, PluginLoader, PluginLocator};
 
 fn create_loader() -> (Sandbox, PluginLoader) {
     let sandbox = create_empty_sandbox();
@@ -27,6 +27,7 @@ mod loader {
                         file: "".into(),
                         path: PathBuf::from("fake-file"),
                     },
+                    &create_http_client().unwrap(),
                 )
                 .await
                 .unwrap();
@@ -44,6 +45,7 @@ mod loader {
                         file: "".into(),
                         path: fixture.join("test.wasm"),
                     },
+                    &create_http_client().unwrap(),
                 )
                 .await
                 .unwrap();
@@ -68,7 +70,9 @@ mod loader {
             loader
                 .load_plugin(
                     Id::raw("test"),
-                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/download/v0.0.2/deno_plugin_invalid_name.wasm".into() })
+                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/download/v0.0.2/deno_plugin_invalid_name.wasm".into() },
+                    &create_http_client().unwrap(),
+                )
                 .await
                 .unwrap();
         }
@@ -80,7 +84,9 @@ mod loader {
             let path = loader
                 .load_plugin(
                     Id::raw("test"),
-                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/download/v0.0.2/deno_plugin.wasm".into() })
+                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/download/v0.0.2/deno_plugin.wasm".into() },
+                    &create_http_client().unwrap(),
+                )
                 .await
                 .unwrap();
 
@@ -94,7 +100,9 @@ mod loader {
             let path = loader
                 .load_plugin(
                     Id::raw("test"),
-                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/latest/download/deno_plugin.wasm".into() })
+                    PluginLocator::SourceUrl { url: "https://github.com/moonrepo/deno-plugin/releases/latest/download/deno_plugin.wasm".into() },
+                    &create_http_client().unwrap(),
+                )
                 .await
                 .unwrap();
 
@@ -118,6 +126,7 @@ mod loader {
                         repo_slug: "moonrepo/invalid-repo".into(),
                         tag: None,
                     }),
+                    &create_http_client().unwrap(),
                 )
                 .await
                 .unwrap();
@@ -135,6 +144,7 @@ mod loader {
                         repo_slug: "moonrepo/bun-plugin".into(),
                         tag: Some("v0.0.3".into()),
                     }),
+                    &create_http_client().unwrap(),
                 )
                 .await
                 .unwrap();
@@ -154,6 +164,7 @@ mod loader {
                         repo_slug: "moonrepo/bun-plugin".into(),
                         tag: None,
                     }),
+                    &create_http_client().unwrap(),
                 )
                 .await
                 .unwrap();